requirements.txt
beautifulsoup4==4.11.2
fastapi==0.115.3
ffmpeg_python==0.2.0
pydantic==2.9.2
Requests==2.32.3
selenium==4.25.0
uvicorn==0.32.0
webdriver_manager==4.0.2
azure-cognitiveservices-speech==1.33.0
openai>=1.28.0
